requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
rembg==2.0.59
Pillow==11.0.0
colorthief==0.2.1
python-multipart==0.0.17
psycopg2-binary==2.9.10
sqlalchemy==2.0.36
python-dotenv==1.0.1
httpx==0.28.0
aiofiles==24.1.0
replicate==0.34.1
